# Conflicts:
#	sculture-web/src/main/java/com/sculture/MainServlet.java
#	sculture-web/src/main/java/com/sculture/controller/GetStory.java
#	sculture-web/src/main/java/com/sculture/helpers/Story.java
#	sculture-web/src/main/webapp/WEB-INF/web.xml
#	sculture-web/src/main/webapp/view_story.jsp

+++ sculture-android/app/src/main/java/tr/edu/boun/cmpe/sculture/models/response/StoryReportResponse.java
@@ -5,11 +5,11 @@ import org.json.JSONObject;
 
 public class StoryReportResponse {
 
-    public int report_count;
+    public long report_count;
 
     public StoryReportResponse(JSONObject object) {
         try {
-            this.report_count = object.getInt("report_count");
+            this.report_count = object.getLong("report_count");
         } catch (JSONException e) {
             e.printStackTrace();
         }
